Episode1-The Beginning
The year is 2052. Only one bookstore remained in the world, which sells paper-based books. Our hero's name is Joe. He loves the old and slow things. We can say it is his fetish. Maybe that is why he decided to rediscover and use the old world's most popular IT tools - at this point we can declare we have a very interesting hero -. He was the most impatient man in the world, that is why he started with the best one, with the EXCEL. The slowest and the most insecure - maybe with this sentence the author overreact the feature of the excel, but he is using it on daily basis and I hope that is enough explanation to you. - tool from the past. Bless God he had a friend, who we can call Krad. He was able to write a macro, which can filter regarding the actually set up and automatically create a form, which contains the result. So Krad finished the work and give it to Joe. Joe promptly started to use, but he couldn't see what after this came...